+++ src/main/java/com/kalsym/product/service/model/store/StoreAsset.java
@@ -26,5 +26,5 @@ public class StoreAsset implements Serializable {
 
     private String logoUrl;
     private String bannerUrl;
-
+    private String bannerMobileUrl;
 }