+++ src/main/java/com/kalsym/product/service/repository/StoreRepository.java
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Repository;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.Query;
 
 /**
  *
@@ -26,5 +27,8 @@ public interface StoreRepository extends PagingAndSortingRepository<Store, Strin
     
     List<Store> findByClientId(@Param("clientId") String clientId);
     
+    @Query("SELECT c FROM Store c WHERE c.isSnooze=true AND c.snoozeEndTime < NOW()")
+    public List<Store> getSnoozeExpired(
+    );
 
 }